Implement stubs and per-job error handlers.
Job stubs effectively reserve space within the scheduler, allowing the user to obtain an (invalid) handle for the job, in case that handle is needed before the job is actually registered. This may be useful for circular references where a Scheduler makes use a JobHandle.

Once the stub is registered with the scheduler, the stub is consumed. If the stub is dropped before being registered the stub and all associated memory is freed without adding anything to the scheduler.

On-error callbacks can now be registered on a per-job basis. This means that reather than having one global handler that is called when _any_ job experiences an error, you can have error handlers attached to specific jobs.