Refocus the lib.
Rota was always meant to be built ontop of tokio and was a required dependency even if the consumer of the library only used the non-async parts.

Maintaining and interacting with blocking and non-blocking versions of everything was clunky and unnecessary.

To that end, everything is async now, all sync code elements have been removed, and naming has been reverted back.

Also trimmed down on some cruft that is no longer needed.