Add convenience method to TaskBuilder for background tasks.
Sometimes it's desirable to have a long-running task in the scheduler that starts and ends when the scheduler does, running continuously.

To that end, TaskBuilder::as_background_task() has been added which automatically sets the schedule to run once and is ready immediately so the task is started immediately after being added to the scheduler and won't be re-started if cancelled.